feat(ios): first working simulator build — real retrieval, visual QA loop
Fixes found by the build+screenshot loop: MLModel Sendable and MainActor
isolation; ON AIR badge clipped by the toolbar's own glass capsule; fp16
mlprogram returns all-zero embeddings on the iOS simulator (fp32 now, model
runs once per question); web-era cluster spans blew the FM 4k budget so
excerpts are capped per-source (RetrievedChunk.capped, tested) instead of
dropping whole sources. Adds -AutoAsk/-SkipOnboarding launch args for
scripted screenshots. 33 kit tests green.

+++ ios/scripts/convert-minilm.py
@@ -80,7 +80,10 @@ def main():
         ],
         outputs=[ct.TensorType(name="embedding")],
         convert_to="mlprogram",
-        compute_precision=ct.precision.FLOAT16,
+        # FLOAT16 produced all-zero outputs on the iOS simulator (fine on
+        # macOS/device); fp32 is ~90 MB and the model runs once per question,
+        # so correctness everywhere beats the size win.
+        compute_precision=ct.precision.FLOAT32,
         minimum_deployment_target=ct.target.iOS17,
     )
     mlmodel.short_description = "all-MiniLM-L6-v2 sentence embedder (mean-pooled, L2-normalized)"
